♻️ Refactor: 중복 정의된 타입 통합 사용 : ProjectResponseType
